Consolidate test cases
