+++ phpstan.neon.dist
@@ -11,7 +11,6 @@ parameters:
 		- src/Views/*
 	ignoreErrors:
 		- '#Call to an undefined static method Config\\Services::[A-Za-z]+\(\)#'
-		- '#Cannot access property [\$a-z_]+ on (array|object)#'
 	universalObjectCratesClasses:
 		- CodeIgniter\Entity
 		- Faker\Generator
