+++ tests/unit/ShieldFactoryTest.php
@@ -1,30 +1,21 @@
 <?php
 
-use CodeIgniter\Test\CIDatabaseTestCase;
 use Sparks\Shield\Models\UserModel;
 use Tatter\Users\Factories\ShieldFactory;
 use Tatter\Users\UserEntity;
+use Tests\Support\FactoryTestCase;
 
-class ShieldFactoryTest extends CIDatabaseTestCase
+class ShieldFactoryTest extends FactoryTestCase
 {
-	/**
-	 * The namespace(s) to help us find the migration classes.
-	 * Empty is equivalent to running `spark migrate -all`.
-	 * Note that running "all" runs migrations in date order,
-	 * but specifying namespaces runs them in namespace order (then date)
-	 *
-	 * @var string|array|null
-	 */
 	protected $namespace = 'Sparks\Shield';
+	protected $class     = ShieldFactory::class;
+	protected $faker     = UserModel::class;
 
-	public function testFactoryMethods()
+	public function testId()
 	{
-		$user = fake(UserModel::class);
-
-		$factory = new ShieldFactory();
-		$result  = $factory->findById($user->id);
+		$result = $this->factory->findById($this->user->id);
 
 		$this->assertInstanceof(UserEntity::class, $result);
-		$this->assertEquals($user->username, $result->getUsername());
+		$this->assertEquals($this->user->username, $result->getUsername());
 	}
 }
